Sooooo I think i made the onboarding logic good overall. The viewModel has livedata that checks if wee recieved the setting permission and the viewPagerFragment has an observer on that livedata. If the livedata is true -> jump to screen number 3.
I will need to make suree all the code is in the right place for the logic, but basically, this is the jist of it.

HOWEVER, i somehow along the way placed a UI BREAKING BUG.
Need to find it and correct it. because right now the application ust breaks the phone.

Next step is to eliminiate components one at a time to find the bug